pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mass-town"
version = "0.1.0"
description = "Prototype workflow supervision system for brittle engineering pipelines."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
mass-town = "mass_town.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/mass_town"]
